options_analysis_app/requirements.txt
requests
pandas
numpy
matplotlib
yfinance
beautifulsoup4
alpha-vantage
scipy
streamlit
finnhub-python
py_vollib
polygon-python-client 